Move database to its own folder. Fix incorrection alembic migration.
