+++ app/database/base.py
@@ -1,0 +1,6 @@
+from sqlalchemy.ext.declarative import declarative_base
+
+Base = declarative_base()
+
+
+from app.data_sources.strong.models import *
